All changes done by Vishal in Jan

+++ webapp/pages/pipelines/[pipelineId].tsx
@@ -102,7 +102,7 @@ const PipelineInfo = ({ pipelineId }: PipelineInfoProps) => {
   if (pipeline === null) return <DefaultErrorPage statusCode={404} />;
   return (
     <Layout
-      title={renderTitle(pipeline, router, setPipeline, setReloadKey)}
+      title={renderTitle(pipeline, router, setPipeline, setReloadKey, setPipelineRuns)}
       subTitle={undefined}
       pageTitle={pipeline ? "Pipeline " + pipeline.name : ""}
       rightBtn={{
@@ -176,7 +176,8 @@ function renderTitle(
   pipeline: PipelineResponseDto | undefined,
   router: NextRouter,
   setPipeline: (value: any) => void,
-  setReloadKey: (value: any) => void
+  setReloadKey: (value: any) => void,
+  setPipelineRuns: (value: any) => void
 ) {
   if (!pipeline) return "";
   const isActive = pipeline.syncStatus === PipelineSyncStatus.ACTIVE;
@@ -227,33 +228,48 @@ function renderTitle(
           }}
         />
       </OverlayTrigger>
-      <Dropdown align="end" className="d-inline-block float-end me-4">
-        <Dropdown.Toggle as={DropdownPlain} id="dropdown-custom-components">
-          <IconDots />
-        </Dropdown.Toggle>
-        <Dropdown.Menu align="end">
-          <Dropdown.Item
-            onClick={() => {
-              pipelineService.restart(pipeline.id).then(() => {
-                setReloadKey(Math.random());
-                bannerNotificationService.success("Pipeline Restarted");
+      <div className="float-end">
+        <Dropdown className="d-inline-block mx-2">
+          <Dropdown.Toggle as={DropdownPlain} id="dropdown-custom-components">
+            <IconDots />
+          </Dropdown.Toggle>
+          <Dropdown.Menu align="end">
+            <Dropdown.Item
+              onClick={() => {
+                pipelineService.restart(pipeline.id).then(() => {
+                  setReloadKey(Math.random());
+                  bannerNotificationService.success("Pipeline Restarted");
+                });
+              }}
+            >
+              Restart
+            </Dropdown.Item>
+            <Dropdown.Item
+              onClick={() => {
+                if (confirm(`Do you want to delete ${pipeline.name} Pipeline ?`)) {
+                  pipelineService.delete(pipeline.id).then(() => {
+                    bannerNotificationService.success("Pipeline Deleted");
+                    router.push("/pipelines").then();
+                  });
+                }
+              }}
+            >
+              Delete
+            </Dropdown.Item>
+          </Dropdown.Menu>
+        </Dropdown>
+        <button className="btn btn-outline-primary mx-2"
+          onClick={() => {
+            pipelineRunsService
+              .getByPipelineId(pipeline.id)
+              .then(({ data }) => {
+                setPipelineRuns(data);
+              })
+              .catch(() => {
+                setPipelineRuns(null);
               });
-            }}
-          >
-            Restart
-          </Dropdown.Item>
-          <Dropdown.Item
-            onClick={() => {
-              pipelineService.delete(pipeline.id).then(() => {
-                bannerNotificationService.success("Pipeline Deleted");
-                router.push("/pipelines").then();
-              });
-            }}
-          >
-            Delete
-          </Dropdown.Item>
-        </Dropdown.Menu>
-      </Dropdown>
+          }}>Refresh</button>
+      </div>
     </>
   );
 }
